Refactor: Move calculations to respective components
